requirements.txt
streamlit
openai
autogen-agentchat
autogen-ext[openai]
google-generativeai